test different params for N

+++ docker/run-thrymr.sh
@@ -1,9 +1,21 @@
 #!/bin/bash
 
+RUN_START=`date`
+rm -f /tmp/stop
+
 cd ~/SWARM/crazys
 
 
 #                                                                                                                         hash                         launch mode   params        obs path extra
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist15 dist mpc1_params2a 0 4 "eel" &
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist9 dist mpc1_params2a 0 4 "eel" &
+sleep 150 # delay compilation by 150 seconds in second two docker containers
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist5 dist mpc1_params2a 0 4 "eel" &
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist1 dist mpc1_params2a 0 4 "eel" &
+sleep 150 # delay compilation by 150 seconds in second two docker containers
+wait
+test -e /tmp/stop && exit 7 # stop if flag is present
+
 docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist15 dist mpc1_params2b 0 4 "eel" &
 docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist9 dist mpc1_params2b 0 4 "eel" &
 sleep 150 # delay compilation by 150 seconds in second two docker containers
@@ -11,6 +23,17 @@ docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyfl
 docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist1 dist mpc1_params2b 0 4 "eel" &
 sleep 150 # delay compilation by 150 seconds in second two docker containers
 wait
+test -e /tmp/stop && exit 7 # stop if flag is present
+
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist15 dist mpc1_params2c 0 4 "eel" &
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist9 dist mpc1_params2c 0 4 "eel" &
+sleep 150 # delay compilation by 150 seconds in second two docker containers
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist5 dist mpc1_params2c 0 4 "eel" &
+docker run --rm --volume ~/SWARM/crazys:/crazyflie_ws/src/crazys crazys /crazyflie_ws/src/crazys/docker/run-simulation.sh `git rev-parse --short HEAD` dist1 dist mpc1_params2c 0 4 "eel" &
+sleep 150 # delay compilation by 150 seconds in second two docker containers
+wait
+test -e /tmp/stop && exit 7 # stop if flag is present
+
 echo -e "\n`date` - done.\n"; exit 0; # --------------- END EXECUTION
